correctly install default ssl socket factory when no key or trust store
